flask/requirements.txt
databricks-sql-connector==4.0.0
databricks-sdk==0.45.0
flask==3.1.1
posit-sdk==0.9.0
